### 2024-10-20 #7349 @Kevin-Mattheus-Moerman
- [ ] In your affiliations, can you spell out USA as `United States of America`, thanks.
- [ ] Your README is very minimal. The following is not a strict requirement, but having separate sections e.g. on getting started, testing, documentation, etc may be good.
- [ ] The references do not appear to render at the end of the paper. I'll ask our team to see if they can spot why that is, but in the meantime, feel free to try to diagnose it yourself too.
