### 2024-10-20 #7349 @Kevin-Mattheus-Moerman
@openjournals/dev can you see why the references are not showing in the paper? Thanks!
