### 2024-10-20 #7349 @yuefan98
Hi @Kevin-Mattheus-Moerman, thanks for bringing these to my attention.

1. Sure, I will change that accordingly
2. We are trying to keep the README minimal while we have comprehensive documentation linked within the README using Sphinx. Here is the link for your reference.  https://nleispy.readthedocs.io/en/latest/.
3. Sure, I will try to fix it as well
