### 2024-10-20 9de9069
added supports for nonlinear RC with constant phase element
