+++ nleis/nleis_elements_pair.py
@@ -117,7 +117,7 @@ def RC(p, f):
     return Rct / (1 + ω_star*1j)
 
 
-@element(num_params=3, units=['Ohm', 'F', ''])
+@element(num_params=3, units=['Ohm', 'F', '-'])
 def RCn(p, f):
     '''
 
@@ -215,7 +215,7 @@ def RCD(p, f):
     return (Z)
 
 
-@element(num_params=6, units=['Ohms', 'F', 'Ohms', 's', '1/V', ''])
+@element(num_params=6, units=['Ohms', 'F', 'Ohms', 's', '1/V', '-'])
 def RCDn(p, f):
     '''
 
@@ -352,7 +352,7 @@ def RCS(p, f):
     return (Z)
 
 
-@element(num_params=6, units=['Ohms', 'F', 'Ohms', 's', '1/V', ''])
+@element(num_params=6, units=['Ohms', 'F', 'Ohms', 's', '1/V', '-'])
 def RCSn(p, f):
     '''
 
@@ -485,7 +485,7 @@ def TP(p, f):
     return Z
 
 
-@element(num_params=4, units=['Ohms', 'Ohms', 'F', ''])
+@element(num_params=4, units=['Ohms', 'Ohms', 'F', '-'])
 def TPn(p, f):
     """
 
@@ -634,7 +634,7 @@ def TDP(p, f):
     return Z
 
 
-@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', ''])
+@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', '-'])
 def TDPn(p, f):
     """
 
@@ -803,7 +803,7 @@ def TDS(p, f):
     return Z
 
 
-@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', ''])
+@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', '-'])
 def TDSn(p, f):
     """
 
@@ -979,7 +979,7 @@ def TDC(p, f):
     return Z
 
 
-@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', ''])
+@element(num_params=7, units=['Ohms', 'Ohms', 'F', 'Ohms', 's', '1/V', '-'])
 def TDCn(p, f):
     """
 
@@ -1107,7 +1107,7 @@ def TDCn(p, f):
 # TLM Model #
 
 
-@element(num_params=6, units=['Ohm', 'Ohm', 'F', 'Ohm', 'F', ''])
+@element(num_params=6, units=['Ohm', 'Ohm', 'F', 'Ohm', 'F', '-'])
 def TLM(p, f):
     """
 
@@ -1157,7 +1157,7 @@ def TLM(p, f):
 ###
 
 
-@element(num_params=8, units=['Ohm', 'Ohm', 'F', '', 'Ohm', 'F', '', ''])
+@element(num_params=8, units=['Ohm', 'Ohm', 'F', '-', 'Ohm', 'F', '-', '-'])
 def TLMn(p, f):
     """
 
@@ -1262,7 +1262,7 @@ def TLMn(p, f):
     return (Z)
 
 
-@element(num_params=6, units=['Ohm', 'Ohm', 'F', 'Ohm', 'F', ''])
+@element(num_params=6, units=['Ohm', 'Ohm', 'F', 'Ohm', 'F', '-'])
 def mTi(p, f):
     """
 
@@ -1343,7 +1343,7 @@ def mTi(p, f):
[…]
