### 2024-10-20 00cc4a8
Update joss paper: fixed reference tracking; minor syntax fix to the documentation
