+++ joss-paper/paper.bib
@@ -82,15 +82,13 @@ OCLC: ocn144517368},
     title = {Nonlinear Electrochemical Impedance Spectroscopy for Lithium-Ion Battery Model Parameterization}, 
     journal = {Journal of The Electrochemical Society} }
 
-
-
 @article{ji2023, 
     doi = {10.1149/1945-7111/ad15ca}, 
     url = {https://iopscience.iop.org/article/10.1149/1945-7111/ad15ca}, 
     year = {2023}, 
     volume = {170},issue = {12}, pages = {123511}, 
     author = {Yuefan Ji and Daniel T. Schwartz}, 
-    title = {Second-Harmonic Nonlinear Electrochemical Impedance Spectroscopy: Part I. Analytical Theory and Equivalent Circuit Representations for Planar and Porous Electrodes}, 
+    title = {Second-Harmonic Nonlinear Electrochemical Impedance Spectroscopy: {Part I}. Analytical Theory and Equivalent Circuit Representations for Planar and Porous Electrodes}, 
     journal = {Journal of The Electrochemical Society} }
 
 @article{ji2024, 
