### 2024-10-20 0b8bb45
Update joss paper: fixed reference tracking; minor syntax fix to the documentation (#22)
* fixed reference tracking
* update affiliation
* minor syntax fix to the documentation
