+++ joss-paper/paper.md
@@ -23,27 +23,27 @@ authors:
     affiliation: 1
 
 affiliations:
- - name: Department of Chemical Engineering & Clean Energy Institute, University of Washington, Seattle, WA, USA
+ - name: Department of Chemical Engineering & Clean Energy Institute, University of Washington, Seattle, WA, United States of America
    index: 1
- - name: Independent Researcher, USA
+ - name: Independent Researcher, United States of America
    index: 2
-date: 27 Sept 2024
+date: 20 Oct 2024
 bibliography: paper.bib
 ---
 
-Building on the growing adoption of impedance.py `[@Murbach2020]` as an open-source software tool within the electrochemical impedance spectroscopy (EIS) community, nleis.py is a toolbox for impedance.py that aims to provide an easily accessible tool to perform second harmonic nonlinear EIS (2nd-NLEIS) analysis, with the ability to extend to higher harmonic analysis in the future. The toolbox is designed with impedance.py in mind to minimize the learning curve for users. It inherits the basic functionality of impedance.py, introduces paired linear and 2nd-harmonic nonlinear circuit elements, and enables the simultaneous analysis of EIS and 2nd-NLEIS. With this toolbox, one can choose to individually analyze an EIS or 2nd-NLEIS spectra or perform simultaneous parameter estimation of linear and nonlinear impedance data using an impedance.py workflow. Ultimately, the nleis.py toolbox will be integrated into impedance.py as adoption grows, while maintaining the standalone version of nleis.py as a platform to develop advanced features as the field matures.
+Building on the growing adoption of impedance.py [@Murbach2020] as an open-source software tool within the electrochemical impedance spectroscopy (EIS) community, nleis.py is a toolbox for impedance.py that aims to provide an easily accessible tool to perform second harmonic nonlinear EIS (2nd-NLEIS) analysis, with the ability to extend to higher harmonic analysis in the future. The toolbox is designed with impedance.py in mind to minimize the learning curve for users. It inherits the basic functionality of impedance.py, introduces paired linear and 2nd-harmonic nonlinear circuit elements, and enables the simultaneous analysis of EIS and 2nd-NLEIS. With this toolbox, one can choose to individually analyze an EIS or 2nd-NLEIS spectra or perform simultaneous parameter estimation of linear and nonlinear impedance data using an impedance.py workflow. Ultimately, the nleis.py toolbox will be integrated into impedance.py as adoption grows, while maintaining the standalone version of nleis.py as a platform to develop advanced features as the field matures.
 
 # Background
 
[…]
