### 2024-10-21 #7349 @yuefan98
Hi @Kevin-Mattheus-Moerman, I have resolved issues with references and updated the affiliation. Please let me know if there are any additional adjustments I need to make at this stage. Thanks!
