### 2024-10-24 feaefdc
Documentation updates and bug fixes: * overwrite the .plot function in NLEISCustomCircuit to ensure correct x and y label * improve the docstring clarity for the customized plot_altair function * improve the handling of negative loglikelihood optimization * update environment.yml to restrict the Python version between 3.8 and 3.11 * rerun the documentation to reflect the changes * bump up the version number to prepare for the next release
