+++ nleis/nleis.py
@@ -334,7 +334,6 @@ class EISandNLEIS:
             return False
 
     def predict(self, frequencies, max_f=10, use_initial=False):
-
         """
         Predict EIS and 2nd-NLEIS using
         linear and nonlinear equivalent circuit model.
@@ -492,7 +491,6 @@ class EISandNLEIS:
         return to_print
 
     def extract(self):
-
         """
         Extracts the parameters of
         EIS and 2nd-NLEIS circuits into dictionaries.
@@ -531,7 +529,6 @@ class EISandNLEIS:
 
     def plot(self, ax=None, f_data=None, Z1_data=None, Z2_data=None,
              kind='nyquist', max_f=10, **kwargs):
-
         """
         Visualizes the model and optional data as Nyquist, Bode,
         or Altair (interactive) plots.
@@ -649,27 +646,28 @@ class EISandNLEIS:
                 mask = np.array(f_pred) < max_f
                 f2 = f_data[mask]
                 Z2 = Z2_data[mask]
-                ax[:, 1] = plot_bode(f2, Z2, units='Ohms/A', ls='', marker='s',
+                ax[:, 1] = plot_bode(f2, Z2, units='Ω/A', ls='', marker='s',
                                      axes=ax[:, 1], **kwargs)
             # we don't need the if else statement
             # if we want to enable plot without fit
             # if self._is_fit():
             Z1_fit, Z2_fit = self.predict(f_pred, max_f=max_f)
-            f1 = f_data
-            f2 = f_data[np.array(f_data) < max_f]
+
+            f1 = f_pred
+            f2 = f_pred[np.array(f_pred) < max_f]
 
             ax[:, 0] = plot_bode(f1, Z1_fit, ls='-', marker='o',
                                  axes=ax[:, 0], **kwargs)
-            ax[:, 1] = plot_bode(f2, Z2_fit, units='Ohms/A', ls='-',
+            ax[:, 1] = plot_bode(f2, Z2_fit, units='Ω/A', ls='-',
                                  marker='o',
                                  axes=ax[:, 1], **kwargs)
 
             ax[0, 0].set_ylabel(r'$|Z_{1}(\omega)|$ ' +
-                                '$[{}]$'.format('Ohms'), fontsize=20)
+                                '$[{}]$'.format('Ω'), fontsize=20)
             ax[1, 0].set_ylabel(
                 r'$-\phi_{Z_{1}}(\omega)$ ' + r'$[^o]$', fontsize=20)
             ax[0, 1].set_ylabel(r'$|Z_{2}(\omega)|$ ' +
-                                '$[{}]$'.format('Ohms/A'), fontsize=20)
+                                '$[{}]$'.format('Ω/A'), fontsize=20)
             ax[1, 1].set_ylabel(
                 r'$-\phi_{Z_{2}}(\omega)$ ' + r'$[^o]$', fontsize=20)
             ax[0, 0].legend(['Data', 'Fit'], fontsize=20)
@@ -716,7 +714,6 @@ class EISandNLEIS:
                              f"'nyquist', or 'bode' (received {kind})")
 
     def save(self, filepath):
-
         """
         Exports the model to a JSON file.
 
@@ -772,7 +769,6 @@ class EISandNLEIS:
             json.dump(data_dict, f)
 
     def load(self, filepath, fitted_as_initial=False):
-
         """
         Imports a model from a JSON file.
 
@@ -956,17 +952,22 @@ class NLEISCustomCircuit(BaseCircuit):
 
         return self
 
-    def predict(self, frequencies, use_initial=False):
+    def predict(self, frequencies, max_f=10, use_initial=False):
         """
 
         Predict impedance using an nonlinear equivalent circuit model
 
         Parameters
         ----------
[…]
