### 2024-10-24 dd0bcde
* Minor improvement in plot handling * rerun the documentation * Improve test coverage
