### 2024-10-24 #7349 @lucydot
> I have resolved issues with references and updated the affiliation.

Confirmed, thank you.
