### 2024-10-25 91daece
Documentation updates and code improvements (#24)
* added tooltips support for the Altair plot
* fixed automatic sorting issue with the Altair
* regenerate the documentation
