### 2024-10-25 #7349 @yuefan98
Hi @lucydot.

Thank you for editing my submission.  Here is a list of potential reviewers who have already agreed to review for JOSS with relevant expertise.

* larsenkg
* alberto-battistel
* leeberty84
* BradyPlanden
* DEARLIBS
* yaomz16
* ericfell
* mefuller
* sudarshanv01
