### 2024-10-28 #7349 @alberto-battistel
> @larsenkg @alberto-battistel @leeberty84 - hello 👋 - are you available to review this JOSS submission?

I am also available.
