### 2024-10-29 7942a3a
Bug fixes * Improve the handling of parameter normalization in simul_fit * Improve the handling of constant parsing * boost test coverage
