### 2024-10-29 34261e5
Bug fixes (#25)
* Improve the handling of parameter normalization in simul_fit
* Improve the handling of constant parsing
* boost test coverage
