### 2024-11-26 0b7afd6
draft data processing function for autolab equipment
