+++ nleis/data_processing.py
@@ -1,0 +1,340 @@
+import numpy as np
+import pandas as pd
+from scipy.fft import rfft, rfftfreq
+from scipy.optimize import curve_fit
+import matplotlib.pyplot as plt
+
+
+def data_loader(filename, equipment='autolab', fft='instrument',
+                multi_current=False, rtol=5e-4, scaling_factor=1000,
+                fit_visualize=False, freq_domain_visual=False):
+    """
+    Load data and perform FFT processing for EIS and 2nd-NLEIS data.
+
+    Parameters:
+    -----------
+    filename : str or list of str
+        Path(s) to the CSV file(s) containing the raw data.
+        list of str is required if using multi_current mode.
+    equipment : str, default 'autolab'
+        Specifies the equipment used for data acquisition.
+    fft : str, default 'instrument'
+        Method used for FFT. Can be 'scipy' or 'instrument'.
+    multi_current : bool, default False
+        If True, process multiple current files.
+    rtol : float, default 5e-4
+        Relative tolerance for finding the frequency components.
+    scaling_factor : int, default 1000
+        Scaling factor for visualizing 2nd-NLEIS results.
+    fit_visualize : bool, default False
+        If True, generate plots to visualize the impedance fit
+        to multiple current.
+        Only available under multi_current mode
+    freq_domain_visual: bool, default False
+        If True, generate frequency domain current and voltage responses
+        at each frequency for the first three harmonics.
+        Only available when multi_current = False
+
+
+    Returns:
+    --------
+    frequencies : ndarray
+        Array of frequency values.
+    Z1 : ndarray
+        Impedance values for EIS.
+    Z2 : ndarray
+        Impedance values for 2nd-NLEIS.
+    """
+    if equipment == 'autolab':
+        if multi_current:
+            if not isinstance(filename, list):
+                raise ValueError(
+                    "filename should be a list of" +
+                    " file paths when multi_current=True")
+
+            df_I1, df_V1, df_V2 = pd.DataFrame(), pd.DataFrame(), \
+                pd.DataFrame()
+
+            for index, file in enumerate(filename):
+                df = pd.read_csv(file)
+                results = autolab_fft(df, method=fft, rtol=rtol)
+                df_I1[index] = results['I1,[A]']
+                df_V1[index] = results['V1,[V]']
+                df_V2[index] = results['V2,[V]']
+
+            frequencies = df['Frequency (Hz)'].dropna()
+            Z1, Z2 = [], []
+
+            for i in range(len(frequencies)):
+                I1 = df_I1.iloc[i].to_numpy().astype(np.complex128)
+                V1 = df_V1.iloc[i].to_numpy().astype(np.complex128)
+                V2 = df_V2.iloc[i].to_numpy().astype(np.complex128)
+
+                Z1_popt, _ = curve_fit(fit_Z1, np.vstack(
+                    [I1.real, I1.imag]), np.hstack([V1.real, V1.imag]))
+                Z2_popt, _ = curve_fit(fit_Z2, np.vstack(
+                    [I1.real, I1.imag]), np.hstack([V2.real, V2.imag]))
+
+                Z1.append(Z1_popt[0] + 1j * Z1_popt[1])
+                Z2.append(Z2_popt[0] + 1j * Z2_popt[1])
+
+                V1_cal, V2_cal = I1 * \
+                    Z1[-1], (I1 ** 2) * Z2[-1] * scaling_factor
+
[…]
