### 2024-11-27 10bc851
initial development of LinMM and NL_MM wit corresponding pytest
