+++ nleis/validation.py
@@ -1,0 +1,317 @@
+import matplotlib.pyplot as plt
+import numpy as np
+from impedance.models.circuits import CustomCircuit
+from .nleis import NLEISCustomCircuit
+from .visualization import plot_first, plot_second
+from tqdm import tqdm
+
+
+def NL_MM(f, Z, raw_circuit='Kn', initial_guess=[0.01, 1], tol=1e-5,
+          plot=False, max_f=10, max_M=20):
+    """
+    Perform NLEIS fitting using nonlinear measurement models
+    with a specified maximum number of elements.
+
+    Parameters:
+    f : array-like
+        Frequency data
+    Z : array-like
+        2nd-NLEIS Impedance data
+    raw_circuit: str
+        Basis function for the measurement model
+    initial_guess : list
+        Initial guess for model parameters
+    tol : float
+        Tolerance for cost convergence
+        Default: 1e-5
+    plot : bool
+        Whether to plot the results during the fitting process
+        Default = False
+    max_f : float
+        Maximum frequency cutoff
+        Default = 10 Hz
+    max_M : int
+        Maximum number of elements to try in the measurement model
+        Default = 20
+
+    Returns:
+    M : int
+        The number of elements used in the circuit model
+    cost : list
+        List of cost values for each iteration
+    p : list
+        Final optimized parameters
+    Z_fit : array
+        Fitted impedance values
+    res_real : array
+        Residuals for the real part of the impedance
+    res_imag : array
+        Residuals for the imaginary part of the impedance
+    conf : array
+        confidence interval of the fitted paramer calculated from covariance
+
+    """
+    if max_M < 1:
+        raise ValueError(
+            'The maximum number of elements (max_M) '
+            + 'should always be greater than or equal to 1')
+
+    circuit = raw_circuit+'0'
+    p0 = initial_guess * max_M
+    n = len(initial_guess)
+    cost = [0]
+    previous_cost = 0
+    mask = f < max_f
+    f = f[mask]
+    Z = Z[mask]
+    with tqdm(total=max_M, file=None) as pbar:
+        for i in range(max_M):
+            M = i + 1
+            model = NLEISCustomCircuit(circuit, initial_guess=p0[:n*M])
+            model.fit(f, Z, max_f=max_f)
+            p0[:n*M] = model.parameters_
+
+            Z_fit = model.predict(f)
+
+            current_cost = cost_max_norm(Z, Z_fit)
+            cost.append(current_cost)
+
+            if plot:
+                fig, ax = plt.subplots(figsize=(5, 5))
+                plot_second(ax, Z)
+                plot_second(ax, Z_fit)
+                ax.text(0.75, 0.2, '# elements = '
+                        + str(i+1),
+                        horizontalalignment='center',
+                        verticalalignment='center',
+                        transform=ax.transAxes, size=12)
+                ax.text(0.75, 0.1, '$COST = $'
+                        + '{:0.2e}'.format(current_cost),
+                        horizontalalignment='center',
+                        verticalalignment='center',
+                        transform=ax.transAxes, size=12)
+                plt.legend(['Data', 'Fit'])
+                plt.show()
+
+            if abs(current_cost - previous_cost) < tol:
+                print(f'Optimal solution found with M = {M}')
+
+                res_real = (Z.real - Z_fit.real) / Z.real
+                res_imag = (Z.imag - Z_fit.imag) / Z.imag
+
+                pbar.update(1)
+                pbar.total = M
+                pbar.refresh()
+
+                p = model.parameters_
+                conf = model.conf_
+                return M, cost[1:], p, Z_fit, res_real, res_imag, conf
+
+            previous_cost = current_cost
+
[…]
