### 2024-12-03 #7424 @leeberty84
> Hi @alberto-battistel and @leeberty84 - how are your reviews going? If you have any questions about the process you can ask here.

Hello Lucydot; I am urgently going back to South Korea from US due to family issue. Can I ask when would be due for this paper? If it should be done in one or two weeks, would it be possible to ask other reviewers to review it (instead of me). Please let me know if you have any concerns. If it is difficult to find other reviewers, I will do my best to review by the due date.
