### 2024-12-09 #7424 @larsenkg
The authors did an excellent job with both the nleis.py repository and the JOSS paper. Installation from source and from PyPI was a breeze, as was running the tests and the example. The paper was well-written and easy to read, with a good overview of the field of EIS/2nd-NLEIS. How nleis.py fits into the ecosystem, with a direct comparison to impedance.py (which currently lacks nonlinear analysis capabilities), was insightful. I don't have any concerns with the paper, the source code, or the documentation.

My only suggestions relates to the README file. I would suggest adding more language identifying the statement of need. This information already exists in the paper and documentation, but is missing from the README. When I first read the README, I was slightly confused by the first sentence, which mentions impedance analysis but does not explicitly mention electrochemical impedance spectroscopy. (My experience is in bioimpedance spectroscopy, so I did not immediately notice the `nl` for nonlinear and `eis` in the package name.) I was not able to understand the target audience until I read the paper. I think one or two more sentences identifying what this project is and who it is for would make it easier for new users to onboard.

My final suggestion is to add a section to the README for new contributors. A potential contributor might miss `CONTRIBUTING.md`. A section that points to this file would be useful. The paper mentions the need for the community to build tools for nonlinear EIS, and encourages researchers in the field to contribute to nleis.py by creating an issue in the repository. Similar language could be added to the README.

Again, great job to the authors!
