### 2024-12-09 #7424 @yuefan98
Hi @larsenkg,

Thank you for your review and compliment about our work. We greatly appreciate your valuable suggestions regarding the README.

To address your feedback, we have :

1. Added a more detailed statement of need to the README to better communicate the project's purpose.
2. Added a new contributing section that directs potential contributors to the `CONTRIBUTING.md` for detailed guidance.

Please let me know if you have any other comments.
