### 2024-12-10 #7424 @larsenkg
Excellent! Those changes address my concerns perfectly.
