### 2024-12-20 #7424 @alberto-battistel
I finished my review.
All the points in the review checklist were satisfied.
It is clear that although works date back more than 30 years there is still no standard way to deal with nonlinear electrochemical impedance spectroscopy. However, this toolbox may fill the gap giving the researcher a common tool to operate.
Of course, there are limitations, like the incomplete development of higher harmonics, nevertheless, this is a great job that will help future works.
