### 2024-12-23 #7424 @yuefan98
Hi @alberto-battistel,

Thanks for your thoughtful and positive review. We truly appreciate your comments about the field and our software. We are also actively working on theories to support the physics-based analysis of both EIS and NLEIS. With our open-source efforts, we aim to continuously support the growth of the field.
