### 2025-01-03 212ea0e
Code Improvement: * add phase correction * add base line correction * general visualization improvement * bug fixes
Unit test:
* Add unit tests for data processing functions in test_data_processing.py
* waiting to add corresponding data
