+++ nleis/nleis_tests/test_data_processing.py
@@ -1,0 +1,178 @@
+import numpy as np
+
+import pytest
+
+import os
+from nleis.data_processing import data_loader, fft_phase_correction, \
+    convert_to_complex, plot_fit, plot_freq_domain
+import pandas as pd
+# import matplotlib
+
+# matplotlib.use("Agg")
+
+test_dir = os.path.dirname(os.path.abspath(__file__))
+data_dir = os.path.join(test_dir, '../data')
+data_path_25mA = os.path.join(data_dir, 'autolab_25mA.txt')
+data_path_50mA = os.path.join(data_dir, 'autolab_50mA.txt')
+data_path_100mA = os.path.join(data_dir, 'autolab_100mA.txt')
+
+data_25mA_raw = data_loader(data_path_25mA, equipment='autolab',
+                            fft='scipy', phase_correction=False,
+                            baseline=False, multi_current=False)
+
+
+def test_data_loader_multi_current():
+
+    # test for multi_current
+    # the data should be a list of data paths
+    # otherwise raise ValueError
+    with pytest.raises(ValueError):
+        data_loader(data_path_25mA, equipment='autolab',
+                    fft='instrument', multi_current=True)
+    # test for multi_current
+    # mismatched data path should lead to the same result
+    path_list_1 = [data_path_25mA, data_path_50mA, data_path_100mA]
+    path_list_2 = [data_path_100mA, data_path_25mA, data_path_50mA]
+    result_list_1 = data_loader(path_list_1, equipment='autolab',
+                                fft='instrument', multi_current=True)
+    result_list_2 = data_loader(path_list_2, equipment='autolab',
+                                fft='instrument', multi_current=True)
+    assert np.allclose(
+        result_list_1[1], result_list_2[1], rtol=1e-2, atol=1e-2)
+
+    # # test to ensure fit_visualize can run without problem
+    # data_loader(path_list_2, equipment='autolab',
+    #             fft='instrument', multi_current=True,
+    #             fit_visualize=True)
+
+
+def test_data_loader_single_current():
+    # test for single current
+    # the data should be a single data path in string
+    # otherwise raise ValueError
+    with pytest.raises(ValueError):
+
+        data_loader([data_path_25mA], equipment='autolab',
+                    fft='instrument', multi_current=False)
+
+    # test for phase correction
+    # the phase corrected impedance
+    # should match non-phase corrected impedance
+    data_25mA_corrected = data_loader(data_path_25mA, equipment='autolab',
+                                      fft='scipy', phase_correction=True,
+                                      baseline=False, multi_current=False)
+    assert np.allclose(
+        data_25mA_raw[1], data_25mA_corrected[1], rtol=1e-2, atol=1e-2)
+
+    # # test to ensure freq_domain_visual can run without problem
+    # data_loader(data_path_25mA, equipment='autolab',
+    #             fft='scipy', phase_correction=True,
+    #             baseline=False, multi_current=False,
+    #             freq_domain_visual=True)
+
+
+def test_data_loader_baseline():
+    # test for baseline correction
+    # the baseline corrected impedance
+    # should match non-baseline corrected impedance for EIS data
+
[…]
