+++ nleis/data_processing.py
@@ -6,10 +6,12 @@ import matplotlib.pyplot as plt
 
 
 def data_loader(filename, equipment='autolab', fft='instrument',
-                multi_current=False, rtol=5e-4, scaling_factor=1000,
+                multi_current=False, rtol=5e-4, phase_correction=True,
+                baseline=True, scaling_factor=1000,
                 fit_visualize=False, freq_domain_visual=False):
     """
     Load data and perform FFT processing for EIS and 2nd-NLEIS data.
+    Currently supports data from Autolab.
 
     Parameters:
     -----------
@@ -21,7 +23,7 @@ def data_loader(filename, equipment='autolab', fft='instrument',
     fft : str, default 'instrument'
         Method used for FFT. Can be 'scipy' or 'instrument'.
     multi_current : bool, default False
-        If True, process multiple current files.
+        If True, process multiple files with different current modulation.
     rtol : float, default 5e-4
         Relative tolerance for finding the frequency components.
     scaling_factor : int, default 1000
@@ -29,7 +31,7 @@ def data_loader(filename, equipment='autolab', fft='instrument',
     fit_visualize : bool, default False
         If True, generate plots to visualize the impedance fit
         to multiple current.
-        Only available under multi_current mode
+        Only available when multi_current = True
     freq_domain_visual: bool, default False
         If True, generate frequency domain current and voltage responses
         at each frequency for the first three harmonics.
@@ -51,20 +53,22 @@ def data_loader(filename, equipment='autolab', fft='instrument',
                 raise ValueError(
                     "filename should be a list of" +
                     " file paths when multi_current=True")
-
+            # initialize DataFrames for storing the current and voltage values
             df_I1, df_V1, df_V2 = pd.DataFrame(), pd.DataFrame(), \
                 pd.DataFrame()
-
+            # process each file in the list
             for index, file in enumerate(filename):
-                df = pd.read_csv(file)
-                results = autolab_fft(df, method=fft, rtol=rtol)
+                df = pd.read_csv(file, low_memory=False)
+                results = autolab_fft(
+                    df, method=fft, rtol=rtol,
+                    phase_correction=phase_correction, baseline=baseline)
                 df_I1[index] = results['I1,[A]']
                 df_V1[index] = results['V1,[V]']
                 df_V2[index] = results['V2,[V]']
 
             frequencies = df['Frequency (Hz)'].dropna()
             Z1, Z2 = [], []
-
+            # Obtain impedance value from curve fitting the current and voltage
             for i in range(len(frequencies)):
                 I1 = df_I1.iloc[i].to_numpy().astype(np.complex128)
                 V1 = df_V1.iloc[i].to_numpy().astype(np.complex128)
@@ -79,11 +83,12 @@ def data_loader(filename, equipment='autolab', fft='instrument',
                 Z2.append(Z2_popt[0] + 1j * Z2_popt[1])
 
                 V1_cal, V2_cal = I1 * \
-                    Z1[-1], (I1 ** 2) * Z2[-1] * scaling_factor
+                    Z1[-1], (I1 ** 2) * Z2[-1]
 
                 if fit_visualize:
                     fig, ax = plot_fit(I1, V1, V2, V1_cal,
                                        V2_cal, scaling_factor)
+                    plt.show()
 
             return frequencies.to_numpy(), np.array(Z1), np.array(Z2)
 
[…]
