### 2025-01-03 cd53555
Code and test improvement: * refactor LinMM ('conf') and NL_MM ('cost') to a single MM function that support 'conf' and 'cost' method * minor bug fixes * updates unit tests to reflect the changes * 100% coverage
