+++ nleis/nleis_tests/test_validation.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from nleis.validation import NL_MM, LinMM, cost_max_norm
+from nleis.validation import MM, cost_max_norm
 from impedance.models.circuits import CustomCircuit
 from nleis import NLEISCustomCircuit
 import pytest
@@ -12,7 +12,7 @@ def suppress_plot_show(monkeypatch):
     monkeypatch.setattr(plt, "show", lambda: None)
 
 
-def test_NL_MM():
+def test_MM_cost_method():
 
     model = NLEISCustomCircuit('RCn', initial_guess=[.1, 1, 0.1])
     f = np.geomspace(1e-2, 1e2, 100)
@@ -20,14 +20,14 @@ def test_NL_MM():
 
     # test invalid max_M raises value error
     with pytest.raises(ValueError):
-        M, cost, p, Z_fit, res_real, res_imag, conf = NL_MM(
-            f, Z2, initial_guess=[.1, 1], tol=1e-5,
-            plot=True, max_f=10, max_M=0)
+        M, cost, p, Z_fit, res_real, res_imag, conf = MM(
+            f, Z2, raw_circuit='Kn', initial_guess=[.1, 1], method='cost',
+            max_f=10, max_M=0, tol=1e-5, plot=True)
 
     # check for an ideal weakly nonlinear RC, M equal to 1
-    M, cost, p, Z_fit, res_real, res_imag, conf = NL_MM(
-        f, Z2, initial_guess=[.1, 1], tol=1e-5,
-        plot=True, max_f=10, max_M=10)
+    M, cost, p, Z_fit, res_real, res_imag, conf = MM(
+        f, Z2, raw_circuit='Kn', initial_guess=[.1, 1], method='cost',
+        max_f=10, max_M=10, tol=1e-5, plot=True)
     assert np.isclose(M, 1)
 
     model = CustomCircuit('TPn0', initial_guess=[1, 1, 1, 0.1])
@@ -35,24 +35,27 @@ def test_NL_MM():
 
     # test optimal solution unfound with max_M raises value error
     with pytest.raises(ValueError):
-        M, cost, p, Z_fit, res_real, res_imag, conf = NL_MM(
-            f, Z2, initial_guess=[.1, 1], tol=1e-5,
-            plot=True, max_f=10, max_M=2)
+        M, cost, p, Z_fit, res_real, res_imag, conf = MM(
+            f, Z2, raw_circuit='Kn', initial_guess=[.1, 1], method='cost',
+            max_f=10, max_M=2, tol=1e-5, plot=True)
 
 
-def test_LinMM():
+def test_MM_conf_method():
     model = CustomCircuit('p(R0,C0)', initial_guess=[.1, 1])
     f = np.geomspace(1e-2, 1e2, 100)
     Z1 = model.predict(f)
 
     # test invalid max_M raises value error
     with pytest.raises(ValueError):
-        M, p, Z_fit, res_real, res_imag, conf = LinMM(
-            f, Z1, initial_guess=[.1, .1], plot=True, CI_plot=True, max_M=0)
+        M, p, Z_fit, res_real, res_imag, conf = MM(
+            f, Z1, raw_circuit='K', initial_guess=[.1, .1], method='conf',
+            max_M=0, plot=True, CI_plot=True)
 
     # check for an ideal RC, M equal to 1
-    M, p, Z_fit, res_real, res_imag, conf = LinMM(
-        f, Z1, initial_guess=[.1, .1], plot=True, CI_plot=True, max_M=10)
+    M, p, Z_fit, res_real, res_imag, conf = MM(
+        f, Z1, raw_circuit='K', initial_guess=[.1, .1], method='conf',
+        max_M=10, plot=True, CI_plot=True)
[…]
