+++ nleis/validation.py
@@ -1,40 +1,127 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from impedance.models.circuits import CustomCircuit
 from .nleis import NLEISCustomCircuit
 from .visualization import plot_first, plot_second
 from tqdm import tqdm
 
 
-def NL_MM(f, Z, raw_circuit='Kn', initial_guess=[0.01, 1], tol=1e-5,
-          plot=False, max_f=10, max_M=20):
-    """
-    Perform NLEIS fitting using nonlinear measurement models
-    with a specified maximum number of elements.
+def MM(f, Z, raw_circuit='K', initial_guess=[0.01, 1], method='cost',
+       max_f=np.inf, max_M=20, tol=1e-5, plot=False, CI_plot=False):
+    '''
+    Wrapper function to perform measurement models fitting with either cost
+    or confidence interval method.
 
     Parameters:
+    -----------
     f : array-like
         Frequency data
     Z : array-like
-        2nd-NLEIS Impedance data
+        Impedance data
     raw_circuit: str
         Basis function for the measurement model
     initial_guess : list
-        Initial guess for model parameters
+        Initial guess for the basis model parameters
+    method : str
+        The method to use for fitting. Either 'cost' or 'conf'
+        Default: 'conf', which reproduces
+        the well-known measurement model program created by Mark Orazem. [1]
+        Convergence is not guaranteed
+        with the 'conf' method.
+        When 'conf' method fails, it is recommended to use the 'cost' method,
+        which is developed us and can be found in ref [2].
+
+    max_f : float
+        Maximum frequency cutoff
+        Default = inf
+    max_M : int
+        Maximum number of elements to try in the measurement model
+        Default = 20
     tol : float
         Tolerance for cost convergence
         Default: 1e-5
     plot : bool
         Whether to plot the results during the fitting process
         Default = False
+    CI_plot : bool
+        Whether to plot the 95% confidence interval calculated using
+        monte carlo simulation with random sampling of 1000 sample
+        using covariance derived standard deviation.
+        Default = False
+
+    Returns:
+    --------
+    M : int
+        The optimal number of elements determined by the fitting
+    cost : list
+        List of cost values for each iteration.
+        Only avaliable when method is 'cost'
+    p : list
+        Final optimized parameters
+    Z_fit : array
+        Fitted impedance values
+    res_real : array
+        Residuals for the real part of the impedance
+    res_imag : array
+        Residuals for the imaginary part of the impedance
+    conf : array
+        confidence interval of the fitted parameter calculated from covariance
+
+    Note:
+    -----
+    When method = 'cost': a total of 7 outputs are returned.
+    When method = 'conf': a total of 6 outputs are returned.
+
+    [1] Orazem, M.E. Measurement model for analysis of
+    electrochemical impedance data.
+    J Solid State Electrochem 28, 1273–1289 (2024).
+    https://doi.org/10.1007/s10008-023-05755-9
+
+    [2] Y. Ji, A. H. Shih D. T. Schwartz
+
+    '''
+
+    if method == 'cost':
+        return MM_cost(f, Z, raw_circuit, initial_guess, max_f, max_M, tol,
+                       plot)
+    elif method == 'conf':
+        return MM_conf(f, Z, raw_circuit, initial_guess, max_f, max_M,
+                       plot, CI_plot)
+    else:
+        raise ValueError('The method should be either cost or conf')
+
+
+def MM_cost(f, Z, raw_circuit='Kn', initial_guess=[0.01, 1],
+            max_f=np.inf, max_M=20, tol=1e-5,
+            plot=False):
+    """
+    Perform NLEIS fitting using nonlinear measurement models
+    with a specified maximum number of elements.
+
+    Parameters:
+    -----------
[…]
