### 2025-01-06 #7424 @yuefan98
Hi @lucydot,

Thanks for initiating the post review checklist. I have completed the post review check.

- [ ] The authors and affiliations (including ORCIDs) are correct
- [ ] The latest release is v0.1.1
- [ ] The Zenodo DOI is 10.5281/zenodo.14606367
- [ ] I have ensured that the title, author list (including ORCIDs), and license listed in the archive match those in the JOSS paper and software license.

Please let me know if there is anything I need to do!
