### 2025-01-08 fa5f0b6
* Add max_k parameter to data_loader for harmonic control * Enable the return of full spectrum harmonic signal for current and voltage
