### 2025-01-10 5f67475
Code Improvement: Vectorized code for fast computation Test Fixes: update expected output in NLEIS tests to reflect correct units
