+++ nleis/nleis_elements_pair.py
@@ -109,7 +109,7 @@ def RC(p, f):
         p[1] = C_{dl}; \\;
 
     """
-    ω = np.array(f)*2*np.pi
+    ω = 2*np.pi * np.array(f)
     Rct, Cdl = p[0], p[1]
 
     ω_star = ω*Rct*Cdl
@@ -154,7 +154,7 @@ def RCn(p, f):
     <https://doi.org/10.1149/1945-7111/ad15ca>`_.
 
     '''
-    ω = np.array(f)*2*np.pi
+    ω = 2*np.pi * np.array(f)
     Rct, Cdl, ε = p[0], p[1], p[2]
 
     ω_star = ω*Rct*Cdl
@@ -206,12 +206,15 @@ def RCD(p, f):
     <https://doi.org/10.1149/1945-7111/ad15ca>`_.
 
     '''
-    ω = np.array(f)*2*np.pi
+    ω = 2*np.pi * np.array(f)
     Rct, Cdl, Aw, τd = p[0], p[1], p[2], p[3]
 
-    Zd = Aw / (np.sqrt(1j*ω*τd) * np.tanh(np.sqrt(1j*ω*τd)))
-    tau = ω*Rct*Cdl
-    Z = Rct / (Rct / (Rct + Zd) + 1j*tau)
+    sqrt_1j_ω_τd = np.sqrt(1j*ω*τd)
+    tanh_1j_ω_τd = np.tanh(sqrt_1j_ω_τd)
+    Zd = Aw / (sqrt_1j_ω_τd * tanh_1j_ω_τd)
+
+    ω_star = ω*Rct*Cdl
+    Z = Rct / (Rct / (Rct + Zd) + 1j*ω_star)
     return (Z)
 
 
@@ -277,22 +280,27 @@ def RCDn(p, f):
 
     '''
 
-    ω = np.array(f)*2*np.pi
+    ω = 2 * np.pi * np.array(f)
     Rct, Cdl, Aw, τd, κ, ε = p[0], p[1], p[2], p[3], p[4], p[5]
 
-    Zd1 = Aw / (np.sqrt(1j*ω*τd) * np.tanh(np.sqrt(1j*ω*τd)))
-    Zd2 = Aw / (np.sqrt(1j*2*ω*τd) * np.tanh(np.sqrt(1j*2*ω*τd)))
+    sqrt_1j_ω_τd = np.sqrt(1j * ω * τd)
+    tanh_1j_ω_τd = np.tanh(sqrt_1j_ω_τd)
+    Zd1 = Aw / (sqrt_1j_ω_τd * tanh_1j_ω_τd)
 
-    ω_star = ω*Rct*Cdl
+    sqrt_1j_2ω_τd = np.sqrt(1j * 2 * ω * τd)
+    tanh_1j_2ω_τd = np.tanh(sqrt_1j_2ω_τd)
+    Zd2 = Aw / (sqrt_1j_2ω_τd * tanh_1j_2ω_τd)
+
+    ω_star = ω * Rct * Cdl
     y1 = Rct / (Zd1 + Rct)
     y2 = Zd1 / (Zd1 + Rct)
 
-    Z1 = Rct / (y1 + 1j*ω_star)
-    const = ((Rct*κ*y2**2) - Rct*ε*F/(R*T)*y1**2) / (Zd2 + Rct)
+    Z1 = Rct / (y1 + 1j * ω_star)
+    const = ((Rct * κ * y2**2) - Rct * ε * F / (R * T) * y1**2) / (Zd2 + Rct)
 
-    Z2 = (const*Z1**2) / (2*ω_star*1j + Rct/(Zd2+Rct))
[…]
