### 2025-01-14 f752d85
Code Improvement: * Enabled circuit processing and circuit evaluation using execution graph * Updated and passed all the corresponding tests * Updated requirements.txt and setup.py * Bumped up version number * Fixed some docstrings
