+++ nleis/fitting.py
@@ -7,6 +7,9 @@ from impedance.models.circuits.elements import circuit_elements
 from impedance.models.circuits.elements import get_element_from_name
 from impedance.models.circuits.fitting import check_and_eval, rmse
 
+import networkx as nx
+import re
+
 # Note: a lot of codes are directly adopted from impedance.py.,
 # which is designed to be enable a easy integration in the future,
 # but now we are keep them separated to ensure the stable performance
@@ -209,6 +212,7 @@ def set_default_bounds(circuit, constants={}):
 
 def circuit_fit(frequencies, impedances, circuit, initial_guess, constants={},
                 bounds=None, weight_by_modulus=False, global_opt=False,
+                graph=False,
                 **kwargs):
     """ Main function for fitting an equivalent circuit to data.
 
@@ -279,6 +283,8 @@ def circuit_fit(frequencies, impedances, circuit, initial_guess, constants={},
     if bounds is None:
         bounds = set_default_bounds(circuit, constants=constants)
 
+    cg = CircuitGraph(circuit, constants)
+
     if not global_opt:
         if 'maxfev' not in kwargs:
             kwargs['maxfev'] = int(1e5)
@@ -289,10 +295,17 @@ def circuit_fit(frequencies, impedances, circuit, initial_guess, constants={},
         if weight_by_modulus:
             abs_Z = np.abs(Z)
             kwargs['sigma'] = np.hstack([abs_Z, abs_Z])
-
-        popt, pcov = curve_fit(wrapCircuit(circuit, constants), f,
-                               np.hstack([Z.real, Z.imag]),
-                               p0=initial_guess, bounds=bounds, **kwargs)
+        if graph:
+            popt, pcov = curve_fit(cg.compute_long, f,
+                                   np.hstack([Z.real, Z.imag]),
+                                   p0=initial_guess,
+                                   bounds=bounds,
+                                   **kwargs,
+                                   )
+        else:
+            popt, pcov = curve_fit(wrapCircuit(circuit, constants), f,
+                                   np.hstack([Z.real, Z.imag]),
+                                   p0=initial_guess, bounds=bounds, **kwargs)
 
         # Calculate one standard deviation error estimates for fit parameters,
         # defined as the square root of the diagonal of the covariance matrix.
@@ -320,6 +333,9 @@ def circuit_fit(frequencies, impedances, circuit, initial_guess, constants={},
             return rmse(wrapCircuit(circuit, constants)(f, *x),
                         np.hstack([Z.real, Z.imag]))
 
+        def opt_function_graph(x):
+            return rmse(cg(f, *x), np.hstack([Z.real, Z.imag]))
+
         class BasinhoppingBounds(object):
             """ Adapted from the basinhopping documetation
             https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.basinhopping.html
@@ -337,8 +353,12 @@ def circuit_fit(frequencies, impedances, circuit, initial_guess, constants={},
 
         basinhopping_bounds = BasinhoppingBounds(xmin=bounds[0],
                                                  xmax=bounds[1])
-        results = basinhopping(opt_function, x0=initial_guess,
-                               accept_test=basinhopping_bounds, **kwargs)
+        if graph:
+            results = basinhopping(opt_function_graph, x0=initial_guess,
+                                   accept_test=basinhopping_bounds, **kwargs)
+        else:
+            results = basinhopping(opt_function, x0=initial_guess,
+                                   accept_test=basinhopping_bounds, **kwargs)
         popt = results.x
 
         # Calculate perror
@@ -583,3 +603,156 @@ def extract_circuit_elements(circuit):
                 current_element.append(char)
     extracted_elements.append(''.join(current_element))
     return extracted_elements
+
+# Circuit Graph for computation optimization
[…]
