+++ requirements.txt
@@ -10,3 +10,4 @@ scipy>=1.0
 pandas >= 2.0.2
 sphinx_rtd_theme
 impedance >= 1.7.1
+networkx>=2.6.3

[…]
