### 2025-01-14 2429eff
Added more test and fixed docstrings
