+++ nleis/nleis_tests/test_graph.py
@@ -1,0 +1,152 @@
+import numpy as np
+import matplotlib.pyplot as plt  # noqa: F401
+
+import timeit
+
+from nleis.nleis import EISandNLEIS, NLEISCustomCircuit  # noqa: F401
+from nleis.fitting import CircuitGraph
+
+
+def test_graph_NLEISCustomCircuit():
+    circ_str = 'd(TDSn0,TDSn1)'
+    initial_guess = [
+        5e-3, 1e-3, 10, 1e-2, 100, 10, 0.1,
+        # TDS0 + additioal nonlinear parameters
+        1e-3, 1e-3, 1e-3, 1e-2, 1000, 1, 0.01,
+        # TDS1 + additioal nonlinear parameters
+    ]
+    # initialize
+    eval_circuit = NLEISCustomCircuit(
+        circ_str, initial_guess=initial_guess, graph=False)
+    graph_circuit = NLEISCustomCircuit(
+        circ_str, initial_guess=initial_guess, graph=True)
+
+    frequencies = np.geomspace(1e-3, 1e1, 100)
+
+    def time_predict_eval():
+        return eval_circuit.predict(frequencies, max_f=np.inf)
+
+    def time_predict_graph():
+        return graph_circuit.predict(frequencies, max_f=np.inf)
+    eval_time = timeit.timeit(time_predict_eval, number=10)
+    graph_time = timeit.timeit(time_predict_graph, number=10)
+    Z2_eval = eval_circuit.predict(frequencies, max_f=np.inf)
+    Z2_graph = graph_circuit.predict(frequencies, max_f=np.inf)
+    assert np.allclose(Z2_eval, Z2_graph)
+    assert graph_time < eval_time
+    print(f'eval_time: {eval_time}, graph_time: {graph_time}')
+
+    # test the fitting of the graph circuit using curve_fit
+    graph_circuit.fit(frequencies, Z2_graph, max_f=np.inf)
+    assert np.allclose(graph_circuit.predict(
+        frequencies, max_f=np.inf), Z2_graph)
+
+    # test the fitting of the graph circuit using with global_opt = True
+    graph_circuit.fit(frequencies, Z2_graph, max_f=np.inf, global_opt=True)
+    assert np.allclose(graph_circuit.predict(
+        frequencies, max_f=np.inf), Z2_graph)
+
+
+def test_graph_EISandNLEIS():
+    circ_str_1 = 'L0-R0-TDS0-TDS1'
+    circ_str_2 = 'd(TDSn0,TDSn1)'
+    initial_guess = [1e-7, 1e-3,  # L0,RO
+                     5e-3, 1e-3, 10, 1e-2, 100, 10, 0.1,
+                     # TDS0 + additioal nonlinear parameters
+                     1e-3, 1e-3, 1e-3, 1e-2, 1000, 1, 0.01,
+                     # TDS1 + additioal nonlinear parameters
+                     ]
+
+    eval_circuit = EISandNLEIS(
+        circ_str_1, circ_str_2, initial_guess=initial_guess, graph=False)
+    graph_circuit = EISandNLEIS(circ_str_1, circ_str_2,
+                                initial_guess=initial_guess, graph=True)
+
+    frequencies = np.geomspace(1e-3, 1e3, 100)
+
+    def time_predict_eval():
+        return eval_circuit.predict(frequencies)
+
+    def time_predict_graph():
+        return graph_circuit.predict(frequencies)
+    eval_time = timeit.timeit(time_predict_eval, number=10)
+    graph_time = timeit.timeit(time_predict_graph, number=10)
+
+    Z1_eval, Z2_eval = eval_circuit.predict(frequencies, max_f=np.inf)
[…]
