+++ nleis/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.1"
+__version__ = "0.2"
 
 try:
     from .nleis import *  # noqa: F401, F403

[…]
