### 2025-01-14 2457aa1
Documentation and docstring improvements:
* Added speed benchmark for graph-based function evaluation to examples
* Added release notes
* Improved CircuitGraph class docstrings
