+++ docs/source/release-notes.rst
@@ -1,0 +1,45 @@
+====================
+Release Notes
+====================
+
+.. Version 0.2
+.. ---------------------------
+
+
+Version 0.1.1 (2025-01-06)
+---------------------------
+This is the official release for the JOSS paper.
+
+**What's Changed**
+
+- Documentation updates by @dt-schwartz and @yuefan98 
+- Bug fixes by @yuefan98 in https://github.com/yuefan98/nleis.py/pull/25
+
+**Full Changelog**: https://github.com/yuefan98/nleis.py/compare/v0.1...v0.1.1
+
+Version 0.1 (2024-09-26)
+-------------------------
+We are excited to announce the first official release of nleis.py! This release marks a significant step forward for nonlinear impedance analysis and will be submitted to JOSS for peer review.
+ 
+**Key Features:**
+
+- Simultaneous fitting and analysis of EIS and 2nd-NLEIS.
+- Full support for nonlinear equivalent circuit (nECM) modeling and analysis.
+- Various linear and nonlinear circuit element pairs derived from existing literature.
+- Seamless integration with impedance.py for expanded impedance analysis capabilities.
+
+**Improvements:**
+
+- Comprehensive [documentation](https://nleispy.readthedocs.io/en/latest/), including a Getting Started guide and API reference.
+- Improved documentation for supported circuit elements.
+- Improved code handling for better performance and readability. 
+
+**Bug Fixes**
+
+- Initial testing and issue resolution to ensure smooth functionality.
+
+**New Contributors**
+
+- Special thanks to @mdmurbach for joining the team and enhancing the package quality.
+
+**Full Changelog**: https://github.com/yuefan98/nleis.py/commits/v0.1
