### 2025-01-20 669c3ae
Code improvement: * simplify the code for parse_circuit * Update compute methods to return impedance value * add networkx dependency to environment.yml
