### 2025-01-28 #7424 @lucydot
Hi @openjournals/bcm-eics - pinging again in case you missed this
